feat(step2): split string expression class into single file
